requirements.txt
Pillow
python-docx
